tasks/backend/Python/AlekseevKirill/task_1/shawarma_stall/pyproject.toml
[tool.poetry]
name = "shawarma-stall"
version = "0.1.0"
description = ""
authors = ["a1ekseevkiri11 <alekseevkirill30092004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
